[FEAT] Progress commit on the friends page
this friend page scaffold is developed, backend handler is yet to be implemented. The groups section is yet to be implemented.
friends.jinja: constructed the left sidebar and the banner, which the design can be copy overed to other pages. Move some content display and js code from old home page to here.