[FEAT] Progress commit on chat page
the chat page scaffold is developed, the js code refreshfriends is newly added and tailored based on the one in old home page. the function is yet to be finished and may induce bugs if you load into this page. If code roll back is necessary to redevelop, delete the js code and the content in the left side card html element.